backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
langgraph==0.0.40
fpdf2>=2.7.4
tavily-python==0.3.3
python-dotenv==1.0.1
groq==0.4.2
supabase>=2.11.0
gotrue>=2.9.2
httpx<0.28.0
python-jose[cryptography]==3.3.0
beautifulsoup4==4.12.3
requests==2.31.0
markdown>=3.5.0